requirements.txt
# Main Agent framework
crewai>=0.28.0
langchain-openai>=0.1.0

# Data analysis
pandas>=2.0.0
numpy>=1.24.0
statsmodels>=0.14.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Environment and configuration management
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Automated test
pytest>=7.0.0
